+++ blockchain/entity/solution.go
@@ -3,6 +3,6 @@ package entity
 import "github.com/tclchiam/block_n_go/blockchain/chainhash"
 
 type BlockSolution struct {
-	Nonce int
+	Nonce uint64
 	Hash  *chainhash.Hash
 }
